start sample refactoring
