+++ Java/export-api-sample/src/main/java/com/perfecto/reporting/sample/api/ExportAllDataCodeSample.java
@@ -1,32 +1,21 @@
 package com.perfecto.reporting.sample.api;
 
-import com.google.gson.*;
-import org.apache.commons.io.IOUtils;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpRequestBase;
-import org.apache.http.client.utils.URIBuilder;
-import org.apache.http.impl.client.HttpClientBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
 
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.concurrent.TimeUnit;
 
+/**
+ * This sample downloads all the data that is related to a given driver execution ID
+ */
 public class ExportAllDataCodeSample {
-    // The Perfecto Continuous Quality Lab you work with
-    public static final String CQL_NAME = "demo"; // TODO put your Continuous Quality Lab name here
 
-    // See http://developers.perfectomobile.com/display/PD/DigitalZoom+Reporting+Public+API on how to obtain a Security Token
-    private static final String PERFECTO_SECURITY_TOKEN = "MY_CONTINUOUS_QUALITY_LAB_SECURITY_TOKEN"; // TODO put your security token here
-
-
-    public static final String REPORTING_SERVER_URL = "https://" + CQL_NAME + ".reporting.perfectomobile.com";
-    public static final String SECURITY_TOKEN = System.getProperty("security-token", PERFECTO_SECURITY_TOKEN);
+    // ********************************************************************************************
+    // Set your "CQL_NAME" and the "PERFECTO_SECURITY_TOKEN" in the ReportiumExportUtils file.
+    // ********************************************************************************************
 
     public static void main(String[] args) throws Exception {
         // TODO replace with your export root directory
@@ -35,18 +24,35 @@ public class ExportAllDataCodeSample {
         // TODO put your driver execution ID here
         String executionId = "MY_DRIVER_EXECUTION_ID";
 
-        // get all the tests of the execution and stores them as JSON to a file
-        JsonObject testExecutionsJson = retrieveTestExecutions(executionId);
+        // get all the tests of the execution as JSON
+        JsonObject testExecutionsJson = ReportiumExportUtils.retrieveTestExecutions(executionId);
 
         JsonArray testExecutionsArray = testExecutionsJson.getAsJsonArray("resources");
         if (testExecutionsArray.size() == 0) {
             System.out.println("There are no test executions for that driver execution ID");
         } else {
+            int testCounter = 1;
+
+            // store each test's data in a separate folder
             for (JsonElement testExecutionElement : testExecutionsArray) {
                 JsonObject testExecution = testExecutionElement.getAsJsonObject();
                 String testId = testExecution.get("id").getAsString();
 
-                // TODO create a folder for test execution exports
+                Path testFolder = Paths.get(exportRoot.toString(), String.format("%03d", testCounter) + "-" + testId);
+                Files.createDirectory(testFolder);
+
+                // write test's data to a file
+                Path testExecutionsJsonPath = Paths.get(testFolder.toString(), "test-execution-" + testId + ".json");
+                ReportiumExportUtils.writeJsonToFile(testExecutionsJsonPath, testExecutionsJson);
+
+                // get all the commands of a specific test and store them as a JSON to file
+                JsonObject commandsJson = ReportiumExportUtils.retrieveTestCommands(testId);
+                Path commandsJsonPath = Paths.get(testFolder.toString(), "commands-" + testId + ".json");
+                ReportiumExportUtils.writeJsonToFile(commandsJsonPath, commandsJson);
+
+                // get PDF of the specific test and store it to a file
+
+
                 // TODO write test execution to file
                 // TODO retrieve commands
                 // TODO download PDF
@@ -59,44 +65,9 @@ public class ExportAllDataCodeSample {
         }
 
 
-
         Path testExecutionsJsonPath = Paths.get(exportRoot.toString(), "test_executions_export.json");
-        writeJsonToFile(testExecutionsJsonPath, testExecutionsJson);
+        ReportiumExportUtils.writeJsonToFile(testExecutionsJsonPath, testExecutionsJson);
 
 
-    }
-
-    private static void writeJsonToFile(Path path, JsonObject content) {
-    }
-
-    private static JsonObject retrieveTestExecutions(String executionId) throws URISyntaxException, IOException {
-        URIBuilder uriBuilder = new URIBuilder(REPORTING_SERVER_URL + "/export/api/v1/test-executions");
-
-         uriBuilder.addParameter("externalId[0]", executionId);
-
-        HttpGet getExecutions = new HttpGet(uriBuilder.build());
-        addDefaultRequestHeaders(getExecutions);
-        HttpClient httpClient = HttpClientBuilder.create().build();
-
-        HttpResponse getExecutionsResponse = httpClient.execute(getExecutions);
-        JsonObject executions;
-        try (InputStreamReader inputStreamReader = new InputStreamReader(getExecutionsResponse.getEntity().getContent())) {
-            Gson gson = new GsonBuilder().setPrettyPrinting().create();
-            String response = IOUtils.toString(inputStreamReader);
-            try {
-                executions = gson.fromJson(response, JsonObject.class);
-            } catch (JsonSyntaxException e) {
-                throw new RuntimeException("Unable to parse response: " + response);
-            }
-            System.out.println("\nList of test executions response:\n" + gson.toJson(executions));
-        }
-        return executions;
-    }
-
-    private static void addDefaultRequestHeaders(HttpRequestBase request) {
-        if (SECURITY_TOKEN == null || SECURITY_TOKEN.equals("MY_CONTINUOUS_QUALITY_LAB_SECURITY_TOKEN")) {
-            throw new RuntimeException("Invalid security token '" + SECURITY_TOKEN + "'. Please set a security token");
-        }
-        request.addHeader("PERFECTO_AUTHORIZATION", SECURITY_TOKEN);
     }
 }
